small warning fix (hopefully)
git-svn-id: svn+ssh://combine-101/home/svn_bio/trunk@318 450597fd-3c47-0410-84d8-f85f2794411d

+++ probcalc_mp/probcalc_mp.c
@@ -1040,7 +1040,7 @@ int read_probcalc_line(FILE *mappingfile, mapping_t * mapping) {
 							break;
 							
 						case 9:
-							mapping->score = atof(field);
+							mapping->score = atoi(field);
 							break;
 							
 						case 10:
